Refactored the settings module as a pydantic BaseSettings class
Enables easily setting environment variables
